refactor(app): wire typed backend domain state
