+++ crates/app/src/app/operation_results.rs
@@ -1,4 +1,4 @@
-use nats_backend::BackendOperation;
+use nats_backend::{BackendErrorContext, BackendOperation};
 
 use crate::tabs::TabKind;
 use crate::toast::ToastLevel;
@@ -10,21 +10,12 @@ impl EasyNatsApp {
         &mut self,
         connection_id: u64,
         operation: BackendOperation,
-        data: serde_json::Value,
     ) {
         if operation == BackendOperation::Publish {
             self.toasts.push(
                 ToastLevel::Success,
                 format!("Published to {}", self.conn_name(connection_id)),
             );
-            return;
-        }
-
-        if self.apply_stream_operation(connection_id, operation, &data)
-            || self.apply_kv_operation(connection_id, operation, &data)
-            || self.apply_obj_store_operation(connection_id, operation, &data)
-            || self.apply_server_info_operation(connection_id, operation, &data)
-        {
             return;
         }
 
@@ -38,7 +29,7 @@ impl EasyNatsApp {
         backend_id: Option<u64>,
         operation: BackendOperation,
         message: &str,
-        data: Option<&serde_json::Value>,
+        context: Option<&BackendErrorContext>,
     ) {
         if operation == BackendOperation::Request
             && let Some(cid) = connection_id
@@ -91,7 +82,7 @@ impl EasyNatsApp {
         }
 
         if let Some(cid) = connection_id {
-            self.clear_kv_loading_on_error(cid, operation, data);
+            self.clear_kv_loading_on_error(cid, operation, context);
             self.clear_obj_store_loading_on_error(cid, operation);
             self.clear_server_info_loading_on_error(cid, operation);
         }
