+++ crates/app/src/tabs/types.rs
@@ -1,18 +1,22 @@
+mod metrics_state;
+mod tab_kind;
+
+pub use metrics_state::MetricsState;
+pub use tab_kind::{AppTabViewer, TabKind};
+
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::time::{Instant, SystemTime};
 
 use eframe::egui;
-use egui_dock::TabViewer;
-use egui_dock::tab_viewer::OnCloseResponse;
-use nats_backend::MetricsSnapshot;
+use nats_backend::{
+    ConsumerInfo, JetStreamAccountInfoSnapshot, KvBucketInfo, KvHistoryItem, ObjectStoreBucketInfo,
+    ObjectStoreObjectInfo, ServerInfoSnapshot, StreamInfo, StreamMessageInfo,
+};
 
 use crate::format::PayloadFormat;
-use crate::i18n::t;
 use crate::proto::ProtoViewState;
-use crate::schema::{MessageSchemaManager, SchemaSelector, SchemaSourceKind, ValidationPolicy};
+use crate::schema::{SchemaSelector, SchemaSourceKind, ValidationPolicy};
 use crate::theme::ThemeId;
-
-use super::guard::TabGuard;
 
 /// Auto-refresh configuration for periodic data reloading.
 #[derive(Debug)]
@@ -60,11 +64,11 @@ pub enum TabAction {
     OpenConsumerEdit {
         connection_id: u64,
         stream_name: String,
-        consumer_json: serde_json::Value,
+        consumer_info: ConsumerInfo,
     },
     OpenKvBucketEdit {
         connection_id: u64,
-        bucket_json: serde_json::Value,
+        bucket_info: KvBucketInfo,
     },
     OpenKvEntryCreate {
         connection_id: u64,
@@ -492,8 +496,8 @@ impl SubscriberState {
 
 #[derive(Debug)]
 pub struct StreamState {
-    pub info: Option<serde_json::Value>,
-    pub messages: Vec<serde_json::Value>,
+    pub info: Option<StreamInfo>,
+    pub messages: Vec<StreamMessageInfo>,
     pub selected_msg: Option<usize>,
     pub payload_format: PayloadFormat,
     pub start_seq: String,
@@ -502,11 +506,11 @@ pub struct StreamState {
     pub batch_size: String,
     pub fetching: bool,
     pub purge_subject: String,
-    pub consumers: Vec<serde_json::Value>,
+    pub consumers: Vec<ConsumerInfo>,
     pub consumers_fetching: bool,
     pub auto_refresh: AutoRefresh,
     pub proto_view: ProtoViewState,
-    pub consumer_fetched: std::collections::HashMap<String, Vec<serde_json::Value>>,
+    pub consumer_fetched: std::collections::HashMap<String, Vec<StreamMessageInfo>>,
     pub consumer_fetching: std::collections::HashSet<String>,
     pub search: ScopedSearchState,
     pub search_generation: u64,
@@ -541,7 +545,7 @@ impl Default for StreamState {
 
 #[derive(Debug)]
 pub struct KvBucketState {
-    pub info: Option<serde_json::Value>,
+    pub info: Option<KvBucketInfo>,
     pub keys: Vec<String>,
     pub selected_key: Option<String>,
     pub search: ScopedSearchState,
@@ -562,7 +566,7 @@ pub struct KvBucketState {
     pub entry_operation: Option<String>,
     pub entry_created: Option<String>,
     pub editor_format: PayloadFormat,
-    pub history: Vec<serde_json::Value>,
+    pub history: Vec<KvHistoryItem>,
     pub history_format: PayloadFormat,
     pub show_history: bool,
     pub auto_refresh: AutoRefresh,
@@ -606,8 +610,8 @@ impl Default for KvBucketState {
 
 #[derive(Debug, Default)]
 pub struct ObjectStoreBucketState {
-    pub info: Option<serde_json::Value>,
-    pub objects: Vec<serde_json::Value>,
+    pub info: Option<ObjectStoreBucketInfo>,
+    pub objects: Vec<ObjectStoreObjectInfo>,
     pub selected_object: Option<String>,
     pub object_filter: String,
     pub loading_objects: bool,
@@ -617,461 +621,7 @@ pub struct ObjectStoreBucketState {
 
 #[derive(Debug, Default)]
 pub struct ServerInfoState {
-    pub server_info: Option<serde_json::Value>,
-    pub account_info: Option<serde_json::Value>,
+    pub server_info: Option<ServerInfoSnapshot>,
+    pub account_info: Option<JetStreamAccountInfoSnapshot>,
     pub loading: bool,
-}
-
-#[derive(Debug, Default)]
-pub struct MetricsState {
-    endpoint: String,
-    latest_attempt: Option<MetricsSnapshot>,
-    history: VecDeque<MetricsSnapshot>,
-    pub loading: bool,
-    pub auto_refresh: AutoRefresh,
-}
-
-impl MetricsState {
-    const MAX_SAMPLES: usize = 720;
-
-    pub fn with_endpoint(endpoint: String) -> Self {
-        let mut state = Self {
-            endpoint,
-            ..Default::default()
-        };
-        state.auto_refresh.enabled = true;
-        state.auto_refresh.interval_secs = 5;
-        state
-    }
-
-    pub fn endpoint(&self) -> &str {
-        self.endpoint.as_str()
-    }
-
-    pub fn endpoint_configured(&self) -> bool {
-        !self.endpoint.trim().is_empty()
-    }
-
-    pub fn set_endpoint(&mut self, endpoint: String) {
-        if self.endpoint == endpoint {
-            return;
-        }
-        self.endpoint = endpoint;
-        self.latest_attempt = None;
-        self.history.clear();
-        self.loading = false;
-        self.auto_refresh.last_refresh = Instant::now();
-    }
-
-    pub fn begin_refresh(&mut self) {
-        self.loading = true;
-        self.auto_refresh.mark_refreshed();
-    }
-
-    pub fn apply_snapshot(&mut self, snapshot: MetricsSnapshot) {
-        self.endpoint = snapshot.endpoint.clone();
-        self.loading = false;
-        if snapshot.has_any_data() {
-            if self.history.len() >= Self::MAX_SAMPLES {
-                self.history.pop_front();
-            }
-            self.history.push_back(snapshot.clone());
-        }
-        self.latest_attempt = Some(snapshot);
-    }
-
-    pub fn has_never_loaded(&self) -> bool {
-        self.latest_attempt.is_none() && self.history.is_empty()
-    }
-
-    pub fn latest_attempt(&self) -> Option<&MetricsSnapshot> {
-        self.latest_attempt.as_ref()
-    }
-
-    pub fn latest_data(&self) -> Option<&MetricsSnapshot> {
-        self.latest_attempt
-            .as_ref()
-            .filter(|snapshot| snapshot.has_any_data())
-            .or_else(|| self.history.back())
-    }
-
-    pub fn history(&self) -> &VecDeque<MetricsSnapshot> {
-        &self.history
-    }
-
-    pub fn is_stale(&self) -> bool {
-        self.latest_attempt
-            .as_ref()
-            .is_some_and(|snapshot| !snapshot.has_any_data())
-            && !self.history.is_empty()
-    }
-}
-
-#[derive(Debug)]
-#[allow(dead_code)] // guard fields exist for RAII Drop cleanup, not direct reads
-pub enum TabKind {
-    Welcome,
-    Publisher {
-        connection_id: u64,
-        connection_name: String,
-        guard: TabGuard,
-        backend_id: u64,
-        state: PublisherState,
-    },
-    Subscriber {
-        connection_id: u64,
-        connection_name: String,
-        guard: TabGuard,
-        backend_id: u64,
-        state: SubscriberState,
-    },
-    Stream {
-        connection_id: u64,
-        connection_name: String,
-        stream_name: String,
-        guard: TabGuard,
-        state: StreamState,
-    },
-    KvBucket {
-        connection_id: u64,
-        connection_name: String,
-        bucket_name: String,
-        guard: TabGuard,
-        state: KvBucketState,
-    },
-    ObjectStoreBucket {
-        connection_id: u64,
-        connection_name: String,
-        bucket_name: String,
-        guard: TabGuard,
-        state: ObjectStoreBucketState,
-    },
-    ServerInfo {
-        connection_id: u64,
-        connection_name: String,
-        guard: TabGuard,
-        state: ServerInfoState,
-    },
-    Metrics {
-        connection_id: u64,
-        connection_name: String,
-        state: MetricsState,
-    },
-    SearchWorkspace {
-        state: SearchWorkspaceState,
-    },
-    MessageSchemas {
-        state: MessageSchemasState,
-    },
-    Settings,
-    LogViewer,
-}
-
-impl TabKind {
-    pub fn title(&self) -> String {
-        match self {
-            TabKind::Welcome => t("common.tab_welcome").to_string(),
-            TabKind::Publisher {
-                connection_name,
-                guard,
-                ..
-            } => {
-                if let Some(id) = guard.display_id() {
-                    format!(
-                        "{} #{} ({})",
-                        t("common.tab_publisher"),
-                        id,
-                        connection_name
-                    )
-                } else {
-                    format!("{} ({})", t("common.tab_publisher"), connection_name)
-                }
-            }
-            TabKind::Subscriber {
-                connection_name,
-                guard,
-                ..
-            } => {
-                if let Some(id) = guard.display_id() {
-                    format!(
-                        "{} #{} ({})",
-                        t("common.tab_subscriber"),
-                        id,
-                        connection_name
-                    )
-                } else {
-                    format!("{} ({})", t("common.tab_subscriber"), connection_name)
-                }
-            }
-            TabKind::Stream {
-                connection_name,
-                stream_name,
-                ..
-            } => {
-                format!("{stream_name} ({connection_name})")
-            }
-            TabKind::KvBucket {
-                connection_name,
-                bucket_name,
-                ..
-            } => {
-                format!("{bucket_name} ({connection_name})")
-            }
-            TabKind::ObjectStoreBucket {
-                connection_name,
-                bucket_name,
-                ..
-            } => {
-                format!("{bucket_name} ({connection_name})")
-            }
-            TabKind::ServerInfo {
-                connection_name, ..
-            } => {
-                format!("{} ({})", t("server_info.title"), connection_name)
-            }
-            TabKind::Metrics {
-                connection_name, ..
-            } => {
-                format!("{} ({})", t("common.tab_metrics"), connection_name)
-            }
-            TabKind::SearchWorkspace { .. } => t("common.tab_search_workspace").to_string(),
-            TabKind::MessageSchemas { .. } => t("common.tab_message_schemas").to_string(),
-            TabKind::Settings => t("settings.title").to_string(),
-            TabKind::LogViewer => t("log_viewer.title").to_string(),
-        }
-    }
-
-    /// Structural identity for use in close actions and deduplication.
-    pub fn tab_id(&self) -> egui::Id {
-        match self {
-            TabKind::Welcome => egui::Id::new("tab:welcome"),
-            TabKind::Publisher {
-                connection_id,
-                backend_id,
-                ..
-            } => egui::Id::new(("tab:publisher", *connection_id, *backend_id)),
-            TabKind::Subscriber {
-                connection_id,
-                backend_id,
-                ..
-            } => egui::Id::new(("tab:subscriber", *connection_id, *backend_id)),
-            TabKind::Stream {
-                connection_id,
-                stream_name,
-                ..
-            } => egui::Id::new(("tab:stream", *connection_id, stream_name.as_str())),
-            TabKind::KvBucket {
-                connection_id,
-                bucket_name,
-                ..
-            } => egui::Id::new(("tab:kv", *connection_id, bucket_name.as_str())),
-            TabKind::ObjectStoreBucket {
-                connection_id,
-                bucket_name,
-                ..
-            } => egui::Id::new(("tab:object-store", *connection_id, bucket_name.as_str())),
-            TabKind::ServerInfo { connection_id, .. } => {
-                egui::Id::new(("tab:server-info", *connection_id))
-            }
-            TabKind::Metrics { connection_id, .. } => {
-                egui::Id::new(("tab:metrics", *connection_id))
-            }
-            TabKind::SearchWorkspace { .. } => egui::Id::new("tab:search-workspace"),
-            TabKind::MessageSchemas { .. } => egui::Id::new("tab:message-schemas"),
-            TabKind::Settings => egui::Id::new("tab:settings"),
-            TabKind::LogViewer => egui::Id::new("tab:log-viewer"),
-        }
-    }
-}
-
-pub struct AppTabViewer<'a> {
-    pub backend: &'a nats_backend::BackendHandle,
-    pub actions: &'a mut Vec<TabAction>,
-    pub search_sources: &'a [SearchSourceSnapshot],
-    pub settings: &'a mut crate::settings::AppSettings,
-    pub theme_id: &'a mut ThemeId,
-    pub log_buffer: &'a crate::log_layer::SharedLogBuffer,
-    pub schema_manager: &'a MessageSchemaManager,
-    pub connections: &'a [(u64, String)],
-}
-
-impl TabViewer for AppTabViewer<'_> {
-    type Tab = TabKind;
-
-    fn title(&mut self, tab: &mut Self::Tab) -> egui::WidgetText {
-        tab.title().into()
-    }
-
-    fn id(&mut self, tab: &mut Self::Tab) -> egui::Id {
-        tab.tab_id()
-    }
-
-    fn ui(&mut self, ui: &mut egui::Ui, tab: &mut Self::Tab) {
-        crate::tabs::viewer::render_tab(self, ui, tab);
-    }
-
-    fn context_menu(&mut self, ui: &mut egui::Ui, tab: &mut Self::Tab, _path: egui_dock::NodePath) {
-        let tab_id = self.id(tab);
-        if ui.button(t("common.close_others")).clicked() {
-            self.actions.push(TabAction::CloseOtherTabs {
-                keep_tab_id: tab_id,
-            });
-            ui.close();
-        }
-        if ui.button(t("common.close_all")).clicked() {
-            self.actions.push(TabAction::CloseAllTabs);
-            ui.close();
-        }
-        if ui.button(t("common.close_to_right")).clicked() {
-            self.actions
-                .push(TabAction::CloseTabsToRight { of_tab_id: tab_id });
-            ui.close();
-        }
-    }
-
-    fn is_closeable(&self, tab: &Self::Tab) -> bool {
-        !matches!(tab, TabKind::Welcome)
-    }
-
-    fn allowed_in_windows(&self, tab: &mut Self::Tab) -> bool {
-        !matches!(tab, TabKind::Welcome)
-    }
-
-    fn on_close(&mut self, tab: &mut Self::Tab) -> OnCloseResponse {
-        // Send explicit Unsubscribe to clean up worker state.
-        // TabGuard Drop handles cancellation + display ID recycling automatically.
-        if let TabKind::Subscriber {
-            connection_id,
-            backend_id,
-            state,
-            ..
-        } = tab
-        {
-            for sub in &state.subscriptions {
-                if sub.active {
-                    self.backend
-                        .send(nats_backend::BackendCommand::Unsubscribe {
-                            connection_id: *connection_id,
-                            backend_id: *backend_id,
-                            subject: sub.subject.clone(),
-                        });
-                }
-            }
-        }
-        OnCloseResponse::Close
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::time::{Duration, SystemTime};
-
-    use nats_backend::{MetricsSection, MetricsSectionError, MetricsSnapshot, VarzMetrics};
-
-    use super::MetricsState;
-
-    fn sample_snapshot(endpoint: &str, collected_at: SystemTime, in_msgs: u64) -> MetricsSnapshot {
-        MetricsSnapshot {
-            endpoint: endpoint.to_string(),
-            collected_at,
-            health: None,
-            varz: Some(VarzMetrics {
-                server_name: Some("local".to_string()),
-                server_id: Some("server-1".to_string()),
-                version: Some("2.11.0".to_string()),
-                host: Some("127.0.0.1".to_string()),
-                port: Some(4222),
-                uptime: Some("1m".to_string()),
-                mem_bytes: 1024,
-                cpu_percent: 1.0,
-                connections: 2,
-                total_connections: 3,
-                subscriptions: 4,
-                slow_consumers: 0,
-                in_msgs,
-                out_msgs: in_msgs,
-                in_bytes: 2048,
-                out_bytes: 2048,
-            }),
-            connz: None,
-            jsz: None,
-            errors: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn metrics_state_marks_last_good_sample_stale_after_failed_refresh() {
-        let mut state = MetricsState::with_endpoint("http://localhost:8222".to_string());
-        let first = sample_snapshot("http://localhost:8222", SystemTime::UNIX_EPOCH, 10);
-        state.apply_snapshot(first);
-
-        state.apply_snapshot(MetricsSnapshot {
-            endpoint: "http://localhost:8222".to_string(),
-            collected_at: SystemTime::UNIX_EPOCH + Duration::from_secs(5),
-            health: None,
-            varz: None,
-            connz: None,
-            jsz: None,
-            errors: vec![MetricsSectionError {
-                section: MetricsSection::Varz,
-                message: "timeout".to_string(),
-            }],
-        });
-
-        assert!(state.is_stale());
-        assert_eq!(state.history().len(), 1);
-        assert_eq!(
-            state.latest_data().unwrap().varz.as_ref().unwrap().in_msgs,
-            10
-        );
-    }
-
-    #[test]
-    fn metrics_state_clears_history_when_endpoint_changes() {
-        let mut state = MetricsState::with_endpoint("http://localhost:8222".to_string());
-        state.apply_snapshot(sample_snapshot(
-            "http://localhost:8222",
-            SystemTime::UNIX_EPOCH,
-            10,
-        ));
-
-        state.set_endpoint("http://localhost:9222".to_string());
-
-        assert!(state.history().is_empty());
-        assert!(state.latest_attempt().is_none());
-        assert_eq!(state.endpoint(), "http://localhost:9222");
-    }
-
-    #[test]
-    fn metrics_state_defaults_to_enabled_auto_refresh_every_five_seconds() {
-        let state = MetricsState::with_endpoint("http://localhost:8222".to_string());
-
-        assert!(state.auto_refresh.enabled);
-        assert_eq!(state.auto_refresh.interval_secs, 5);
-    }
-
-    #[test]
-    fn metrics_state_keeps_one_hour_at_default_refresh() {
-        let mut state = MetricsState::with_endpoint("http://localhost:8222".to_string());
-        for idx in 0..=MetricsState::MAX_SAMPLES {
-            state.apply_snapshot(sample_snapshot(
-                "http://localhost:8222",
-                SystemTime::UNIX_EPOCH + Duration::from_secs(idx as u64 * 5),
-                idx as u64,
-            ));
-        }
-
-        assert_eq!(state.history().len(), MetricsState::MAX_SAMPLES);
-        assert_eq!(
-            state
-                .history()
-                .front()
-                .unwrap()
-                .varz
-                .as_ref()
-                .unwrap()
-                .in_msgs,
-            1
-        );
-    }
 }
